pull UITableViewCell stuff into its own file
