+++ SweetDataSourceDemo/SweetDataSource/Extensions/UITableView+Generics.swift
@@ -8,22 +8,6 @@
 
 import UIKit
 
-public extension UITableViewCell {
-
-    /// A default reuse identifier based on the cell's class.
-    public static var defaultReuseIdentifier: String {
-        return String(describing: type(of: self))
-    }
-
-    /// Gets a nib with a name matching that of the cell's class.
-    /// - Parameters:
-    ///    - bundle: The bundle to grab the nib from.
-    public static func defaultNib(inBundle bundle: Bundle = Bundle.main) -> UINib {
-        let nibName = String(describing: type(of: self))
-        return UINib(nibName: nibName, bundle: bundle)
-    }
-}
-
 public extension UITableView {
 
     public func registerCell<CellType: UITableViewCell>(of type: CellType.Type) {
